feat(sqlite): replace monolithic checkpoint module with a dedicated writes submodule
The entire `checkpoint.rs` file has been removed and its contents split into a new `checkpoint/writes.rs` submodule, keeping the SQLite-backed checkpointer implementation but organising the code more cleanly. This change is purely structural with no behavioural impact on the checkpointer's functionality.

Auto-committed-on: dragonfly
